Add test for __mulsi3
This test is a bit different from the others: I've written it in such a
way that it should be run in simavr (instead of on real AVR hardware).
This makes it much easier to test the code without fiddling with
hardware.

To run this test locally, simply run:

    make test

I ran this test for over 100000000 (1e8) pairs of numbers without any
mismatches.